requirements.txt
python-dotenv
# --------------------
# CORE
# --------------------

fastapi[standard] == 0.116.1
pydantic>=2.8.2
sqlalchemy>=2.0.32
alembic>=1.13.2
python-multipart>=0.0.9
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# --------------------
# FILE PROCESSING (CV parsing)
# --------------------
PyMuPDF>=1.24.9          # do PDF
python-docx>=1.1.2


# --------------------
# AI / NLP
# --------------------
transformers>=4.44.2
torch>=2.4.1
spacy>=3.7.5
nltk==3.9.1

# --------------------
# TESTS
# --------------------
pytest>=8.3.2
pytest-asyncio>=0.23.8
httpx>=0.27.0

# --------------------
# DEV TOOLS
# --------------------
black>=24.8.0
isort>=5.13.2
flake8>=7.1.1
pre-commit>=3.8.0